+++ flutter_color_models/lib/src/models/hsb_color.dart
@@ -1,3 +1,4 @@
+import 'dart:ui' as ui;
 import 'package:flutter/painting.dart' show Color;
 import 'package:color_models/color_models.dart' as cm;
 import '../color_model.dart';
@@ -9,6 +10,7 @@ import 'helpers/cast_to_color.dart';
 class HsbColor extends cm.HsbColor
     with AsColor, RgbGetters, CastToColor
     implements ColorModel {
+
   /// {@macro color_models.HsbColor.constructor}
   const HsbColor(
     num hue,
@@ -109,8 +111,38 @@ class HsbColor extends cm.HsbColor
     return copyWith(alpha: (opacity * 255).round());
   }
 
+  // We must override this withValues from Color, has this from dart ui
+  /// Returns a new color that matches this color with the passed in components
+  /// changed.
+  ///
+  /// Changes to color components will be applied before applying changes to the
+  /// color space.
+  Color withValues(
+      {double? alpha,
+      double? red,
+      double? green,
+      double? blue,
+      ui.ColorSpace? colorSpace}) {
+    Color? updatedComponents;
+    if (alpha != null || red != null || green != null || blue != null) {
+      updatedComponents = Color.from(
+          alpha: alpha ?? a,
+          red: red ?? r,
+          green: green ?? g,
+          blue: blue ?? b,
+          colorSpace: this.colorSpace);
+    }
+    if (colorSpace != null && colorSpace != this.colorSpace) {
+      final UICloned_ColorTransform transform =
+          UICloned_getColorTransform(this.colorSpace, colorSpace);
+      return transform.transform(updatedComponents ?? this, colorSpace);
+    } else {
+      return updatedComponents ?? this;
+    }
+  }
+  
   @override
-  HsbColor withValues(List<num> values) {
+  HsbColor withValuesList(List<num> values) {
     assert(values.length == 3 || values.length == 4);
     assert(values[0] >= 0 && values[0] <= 360);
     assert(values[1] >= 0 && values[1] <= 100);
@@ -202,4 +234,25 @@ class HsbColor extends cm.HsbColor
 
   @override
   HsbColor convert(cm.ColorModel other) => other.toHsbColor().cast();
+
+  //OVERRIDEs for painting.dart Color
+  @override
+  double get a => (alpha / 255);
+
+  /// The red channel of this color.
+  @override
+  double get r =>(red / 255);
+
+  /// The green channel of this color.
+  @override
+  double get g => (green / 255);
+
+  /// The blue channel of this color.
+  @override
+  double get b => (blue / 255);
+
+  /// The color space of this color.
+  @override
+  final ui.ColorSpace colorSpace=ui.ColorSpace.sRGB;
+    
 }