+++ flutter_color_models/lib/src/models/oklab_color.dart
@@ -1,3 +1,4 @@
+import 'dart:ui' as ui;
 import 'package:flutter/painting.dart' show Color;
 import 'package:color_models/color_models.dart' as cm;
 import '../color_model.dart';
@@ -99,7 +100,7 @@ class OklabColor extends cm.OklabColor
   @override
   OklabColor withAlpha(int alpha) {
     assert(alpha >= 0 && alpha <= 255);
-    return OklabColor(lightness, a, b, alpha);
+    return OklabColor(lightness, lab_a, lab_b, alpha);
   }
 
   @override
@@ -108,8 +109,38 @@ class OklabColor extends cm.OklabColor
     return copyWith(alpha: (opacity * 255).round());
   }
 
+  // We must override this withValues from Color, has this from dart ui
+  /// Returns a new color that matches this color with the passed in components
+  /// changed.
+  ///
+  /// Changes to color components will be applied before applying changes to the
+  /// color space.
+  Color withValues(
+      {double? alpha,
+      double? red,
+      double? green,
+      double? blue,
+      ui.ColorSpace? colorSpace}) {
+    Color? updatedComponents;
+    if (alpha != null || red != null || green != null || blue != null) {
+      updatedComponents = Color.from(
+          alpha: alpha ?? a,
+          red: red ?? r,
+          green: green ?? g,
+          blue: blue ?? b,
+          colorSpace: this.colorSpace);
+    }
+    if (colorSpace != null && colorSpace != this.colorSpace) {
+      final UICloned_ColorTransform transform =
+          UICloned_getColorTransform(this.colorSpace, colorSpace);
+      return transform.transform(updatedComponents ?? this, colorSpace);
+    } else {
+      return updatedComponents ?? this;
+    }
+  }
+  
   @override
-  OklabColor withValues(List<num> values) {
+  OklabColor withValuesList(List<num> values) {
     assert(values.length == 3 || values.length == 4);
     if (values.length == 4) assert(values[3] >= 0 && values[3] <= 255);
     return OklabColor.fromList(
@@ -124,8 +155,8 @@ class OklabColor extends cm.OklabColor
     assert(alpha == null || (alpha >= 0 && alpha <= 255));
     return OklabColor(
       lightness?.toDouble() ?? this.lightness,
-      a?.toDouble() ?? this.a,
-      b?.toDouble() ?? this.b,
+      a?.toDouble() ?? this.lab_a,
+      b?.toDouble() ?? this.lab_b,
       alpha ?? this.alpha,
     );
   }
@@ -174,4 +205,25 @@ class OklabColor extends cm.OklabColor
 
   @override
   OklabColor convert(cm.ColorModel other) => other.toOklabColor().cast();
+
+  //OVERRIDEs for painting.dart Color
+  @override
+  double get a => (alpha / 255);
+
+  /// The red channel of this color.
+  @override
+  double get r =>(red / 255);
+
+  /// The green channel of this color.
+  @override
+  double get g => (green / 255);
+
+  /// The blue channel of this color.
+  @override
+  double get b => (blue / 255);
+
+  /// The color space of this color.
+  @override
+  final ui.ColorSpace colorSpace=ui.ColorSpace.sRGB;
+    
 }